Add files via upload
This game is very popular and is fairly simple by itself. It is actually a two player game. In this game, there is a board with n x n squares. In this game, it is 3 x 3 squares. The goal of Tic-Tac-Toe is to be one of the players to get three same symbols in a row - horizontally, vertically or diagonally - on a 3 x 3 grid.
This game can be played in a 3x3 grid. The game can be played by two players.
A player can choose between two symbols with his opponent, usual games use “X”and “O”. If first player choose “X” then the second player have to  play with “O” and vice versa. 
A player marks any of the 3x3 squares with his symbol (may be “X” or “O”) and his aim is to create a straight line horizontally or vertically or diagonally with two intensions:
 a) Create a straight line before his opponent to win the game.
  b) Restrict his opponent from creating a straight line first. In case logically no one can create a straight line with his own symbol, the game results a tie.
 Hence there are only three possible results – a player wins, his opponent (human or computer) wins or it’s a tie.

ALGORITHM OF THE PROGRAM

Step 1: Start
Step 2: Decide who goes first and who goes second by checking                      IF((player==player%2)?1:2)
Step 3: Player 1 will mark ‘X’ whereas Player 2 will mark ‘O’
Step 4: After 9 valid moves there will be a result either Player 1 wins or Player 2 or the game will end with a draw by checking IF(i==1)
Step 5: Print the result
Step 6: Stop